Basically complete, functional solitaire.
Main menu lets either play a game or exit

Game loop checks to see if the player wins, and if so, print a special
message and exit to the main loop

All moves fully supported (& mostly tested 🙄)

Mainly wanted to move on to bigger & better things, to wit, although
it's very functional, and I spent more time playing this simple thing
than I care to admit, it's challenging to tell what colours the suits
are which makes it a bit of a pita.  So didn't test through to a win
or a myriad of other things that should be tested.

n.b.: I went fairly out of my way to make this dreadfully bad for
illustrative purposes.  If you're a pointy-hair type evaluating my
work, you have to accept my word for that.  I could, like every other
IT hack, have bumped my clout score inveighing as if I were the King
of France on the topic of object oriented design, or reusability, or
modularity, or proper exception handling, etc., etc., but that is not
what I'm about here.  If you're a typical pointy hair you won't have
been able to maintain an attention span to read this far into the
comments, but if you have, then you are an exceptional specimen of
the field and I expect you're competent to get this gist.