Some cleanup stuff:
+ imported colors & the fini function from new-ui just to make things
a little easier (& funner) for testing.
+ generally, DRY = don't repeat yourself.  If I perform the same
regular expression three separate places, either put it in a function
or, in this case, just run it once and save the results in something
I can reuse later
+ organize the logic by the type of thing I'm operating on: some
things operate on moves, so there are a variety of move functions.
Other things operate on cards, which are now an array of two elements,
the value and the suit.  Rather than using the same logic in 10
places to extract the suit, now I call a function, so I could change
how the card is stored however I want and only a couple functions
need updating
+ it was irritating to have a moveTo function that worked for two
out of three cases but not moving from the board.  So make a little
change to make it consistent and use it in three cases.